Revert "Implement a decision task for taskcluster jobs"
This reverts commit f57ead1924c978dc9256d8f572614731a1a51678.
