+++ tools/tox.ini
@@ -8,11 +8,8 @@ deps =
   pytest-cov
   mock
   hypothesis
+  # `requests` is required by `pr_preview.py`
   requests
-  taskcluster
-  pyyaml
-  json-e
-  jsonschema
 
 commands = pytest {posargs}
