+++ tools/wpt/paths
@@ -3,4 +3,5 @@ tools/docker/
 tools/lint/
 tools/manifest/
 tools/serve/
+tools/taskcluster/
 tools/wpt/
